A more compact and useful representation for unphased data
Previously, unphased data was stored just like phased data. The only
difference was a flag set in the file header, otherwise you still had
haploid sample lists that you need to traverse and count the number
of copies based on the ploidy.

Now we store the number of copies explicitly in the position index,
where we reserve 8-bits for this (so supports ploidy up to 255).
Each variant row then has an associated number of copies, and you
can just skip a variant if it has a number of copies you don't
care about. For example, with diploid data you can do analyses with
just homozygotes by only scanning variants with numCopies=2, which
would massively speed up your analysis.