Remove VCF "rotate" function
As far as I know, this function was unused and just unnecessarily
bloating the picovcf.hpp header. This function was created before
IGD was defined, and now if you want to perform a rotation you
should use IGD, both for performance reasons and for ease of use
of the APIs.

If there is a request/need, I will add examples of how to rotate
an IGD file.