Add metadata export to igdtools (#34)
igdtools now supports exporting variant-based metadata to files
that can be loaded with numpy.loadtxt(). Matrix-based metadata
(i.e., for VCF this means FORMAT fields other than GT) is not
supported: if you need per-variant-per-sample metadata, then there
is probably no reason to use IGD (you need a non-sparse
representation like VCF/BCF).

There are two ways to export this metadata:
1. During VCF->IGD conversion: `igdtools in.vcf.gz -o out.igd -e all`
2. Just export metadata: `igdtools in.vcf.gz -e all`

The metadata is stored as a file per metadata item type. The
supported fields are CHROM, QUAL, FILTER, and INFO. For INFO, each
key gets its own file.  All metadata files are a single entry (line)
per variant in the resulting IGD file (i.e., "expanded" variants).

The first line of a metadata file is a comment that has information
about the metadata. When loaded with numpy.loadtxt(), the size of
the array is exactly igd_file.num_variants in length, and if you
index variant `i` in the IGD file you can get its metadata by
looking at element `i` of the corresponding metadata array.

When a metadata value is not provided for a particular variant,
a default value is used based on the Type field in the VCF metadata:
* Integer: 0
* Float: NaN
* String: .