HTSLIB integrated into picovcf -- not sure if I will use
Integrated HTSLIB as an alternate "backend" for VCF parsing. The
advantage is that it supports tabix (in fact, my implementation
requires tabix), and there is a nice API to just pick the "best"
VCF parser for your file.
Tabix support is useful for parallel VCF parsing/converting. I
particularly want it for converting VCFs to IGD more efficiently.

Downside is that it complicates things a bit, including compile-
time breaking backwards compatibility. Another downside is that it
is actually slower than picovcf for just straight-ahead parsing a
file from beginning to end -- like 50% slower. It is possible that
this is due to the way I am using HTSLIB, but that is unclear.

I'm pushing this branch for now, to save the work. What I expect
will actually end up happening:
1. There are quite a few good API changes on this branch in general
   that I will port to the main branch.
2. I will implement tabix index support for picovcf natively, since
   it appears that picovcf's parser is faster. Doing this will
   result in a simpler changeset that is faster than htslib anyways.

Some of the API changes contained in this branch that are good to
port to main:
* getGenotypeArray() instead of GTIndividualIterator. Much faster,
  and not limited to haploid/diploid samples. GTIndividualIterator
  can stick around, but be marked deprecated.
* forceToPloidy = X instead of the ploidy handling I had before.
  This is more general/useful.
* Deprecate hasNextVariant() in favor of just returning a bool from
  nextVariant().