Work in progress: Add custom tabix support
Users can read the index and use it directly if desired. The only
usage that picovcf has is for lowerBoundPosition() which binary
searches the linear tabix index to massively speed up the seeking
to positions in the VCF file.

Some of the other API changes contained in this branch that are
good to port to main:
* getGenotypeArray() instead of GTIndividualIterator. Much faster,
  and not limited to haploid/diploid samples. GTIndividualIterator
  can stick around, but be marked deprecated.
* forceToPloidy = X instead of the ploidy handling I had before.
  This is more general/useful.
* Deprecate hasNextVariant() in favor of just returning a bool from
  nextVariant().