Contig handling for picovcf and igdtools
You can select from multiple options on how to handle contigs
encountered in a VCFFile. The goal is to make it easy to handle
the scenario where the file has a single contig, and easy to tell
when that is not the case.

Updated documentation reflects how to use these options. IGD only
supports a single contig (i.e., it doesn't have the concept of
sequences within the format), so igdtools has options for how to
handle contigs when encountered converting VCF to IGD.