Tabix index support; parallel igdtools (#45)
* Support for Tabix indexes

Users can read the index and use it directly if desired. The only
usage that picovcf has is for lowerBoundPosition() which binary
searches the linear tabix index to massively speed up the seeking
to positions in the VCF file.

Tabix support lets us seek to different regions of the VCF file
rapidly, which makes parallel processing more reasonable. igdtools
now supports multiple threads when converting VCF.GZ to IGD, or when
exporting metadata from VCF.GZ.

This could be improved by:
1. Instead of parallelizing by BP range, just do based on BGZF
   blocks.
2. Right now the seeking and regular reading use the same buffer
   size for read-ahead (compressed data). I want to try a small
   one for seeking and a large one for regular reading; I suspect
   that blocking read() calls are what is preventing the parallelism
   from exceeding like 5x/6x speedup.

* Update docs, bump version number

Tabix support and parallel igdtools is the last feature for
release v2.4. Updated the documentation to refer to the new
features.