Add handling of "." as matching the ploidy for unphased data
For unphased data we allow "." (haploid missing) even when the rest
of the data is not haploid, because this convention is sometimes
used in VCFs. In general, when converting from VCF->IGD if we see
"." as any of the alleles for unphased data then we will set the
entire individual to be missing. IGD only supports all-or-nothing
missingness per individual for unphased data.